feat: add attach to an existing session (#337)
Added a method to attach to an existing session into the driver class so that
users can attach to an existing session for debugging.

The capabilities only has the given automationName and platformName for the internal use. W3C does not define getting a session info, so we need to give them.